fix minor bug in Fill in argument name

+++ plugins/com.redhat.ceylon.eclipse.ui/src/com/redhat/ceylon/eclipse/code/correct/FillInArgumentNameProposal.java
@@ -30,15 +30,28 @@ class FillInArgumentNameProposal extends CorrectionProposal {
             if (e!=null) {
                 final String name = id.getText();
                 if (e.getTerm() instanceof Tree.FunctionArgument) {
+                    //convert anon functions to typed named argument
+                    //i.e.     (Param param) => result;
+                    //becomes  function fun(Param param) => result;
+                    //and      (Param param) { return result; };
+                    //becomes  function fun(Param param) { return result; }
+                    //and      void (Param param) {};
+                    //becomes  void fun(Param param) {}
                     Tree.FunctionArgument fa = (Tree.FunctionArgument) e.getTerm();
                     if (!fa.getParameterLists().isEmpty()) {
                         int startIndex = fa.getParameterLists().get(0).getStartIndex();
                         if (fa.getType().getToken()==null) {
+                            //only really necessary if the anon 
+                            //function has a block instead of => 
                             change.addEdit(new InsertEdit(startIndex, "function "));
                         }
                         change.addEdit(new InsertEdit(startIndex, name));
                         try {
-                            if (doc.getChar(sa.getStopIndex())==';') {
+                            //if it is an anon function with a body,
+                            //we must remove the trailing ; which is
+                            //required by the named arg list syntax
+                            if (fa.getBlock()!=null &&
+                                    doc.getChar(sa.getStopIndex())==';') {
                                 change.addEdit(new DeleteEdit(sa.getStopIndex(), 1));
                             }
                         }
@@ -46,6 +59,9 @@ class FillInArgumentNameProposal extends CorrectionProposal {
                     }
                 }
                 else {
+                    //convert other args to specified named args
+                    //i.e.     arg;
+                    //becomes  name = arg;
                     change.addEdit(new InsertEdit(sa.getStartIndex(), name + " = "));
                 }
                 if (change.getEdit().hasChildren()) {
